fix(plugins) urldecode path fragments which accept spaces
In some plugins (basic-auth, hmac-auth, jwt, key-auth, oauth2)
we have path fragments which represent plugin fields that may
contain spaces in their definitions. This means that, if a
username for basic-auth contains a space, for example, then
it is reasonable to perform this request:

```
http ":8001/consumers/bob/basic-auth/John Doe/"
```

which, due to URL-encoding, results in a request to
http://localhost:8001/consumers/bob/basic-auth/John%20Doe/ --
we are currently returning 404 for those queries.

This patch performs URL-decoding on those path fragments,
reusing the urldecode function that was present in the
aws-lambda plugin.

It also modifies the test suites for these plugins using values
with spaces for the relevant fields, and using the URL-encoded
version of these fields when querying the Admin API.

Fixes #3247.